Osa 4, tehtävät 4.15. - 4.22.

+++ osa4/blogilista/controllers/blogs.js
@@ -1,12 +1,18 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
 blogRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', {
+        username: 1,
+        name: 1
+    })
     response.json(blogs)
 })
 
-blogRouter.post('/', async (request, response) => {
+blogRouter.post('/', userExtractor, async (request, response) => {
     const body = request.body
 
     let likes = 0
@@ -14,14 +20,24 @@ blogRouter.post('/', async (request, response) => {
         likes += body.likes
     }
 
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const user = request.user
+
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: likes
+        likes: likes,
+        user: user._id
     })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
 
     if (!body.title || !body.url) {
         response.status(400).end()
@@ -30,9 +46,24 @@ blogRouter.post('/', async (request, response) => {
     }
 })
 
-blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+blogRouter.delete('/:id', userExtractor, async (request, response) => {
+
+    const blog = await Blog.findById(request.params.id)
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const user = request.user
+
+    if (blog.user && blog.user.toString() === user.id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    } else {
+        return response.status(403).json({ error: 'You are not authorized to delete this blog' });
+    }
+    
 })
   
 blogRouter.put('/:id', async (request, response) => {
